LIBITD-2704. Move admin functionality to "/admin" URL path
Moved administrative functionality from the "/prospects" URL path
(which is also used by students to submit applications) to a separate
"/admin" URL path.

This is intended to enable the admin functions on the "/admin" path
to be limited to users on-campus, or using the campus VPN.

The major changes are:

* Adding an "admin" namespace to "config/routes.rb"
* Moving the admin functionality in the "ProspectsController" into a
  new "Admin::ProspectsController".
* Moving the "ConfigurationController" and "UsersController" to
  "Admin::ConfigurationController" and "Admin::UsersController"
* Moving the admin-related views into "views/admin" subdirectories
* Updating the paths to partials and to URLs as needed to reflect the
  addition of the "admin" namespace. This also required making some
  paths to partials remaining in the "views/prospects" more explicit
  (i.e., "address_fields" to "prospects/address_fields").
* Updated the paths in the tests are needed

https://umd-dit.atlassian.net/browse/LIBITD-2704
