+++ app/controllers/admin/base_controller.rb
@@ -1,0 +1,5 @@
+module Admin
+  class BaseController < ApplicationController
+    before_action :ensure_auth
+  end
+end
