+++ app/views/admin/prospects/_admin_section.html.erb
