+++ app/views/admin/prospects/_availability_modal.html.erb
