+++ app/views/admin/prospects/_deactivate_modal.html.erb
@@ -7,13 +7,13 @@
       </div>
       <div class="modal-body">
           <table class="table deactivate-table">
-            <tbody></tbody> 
-          </table> 
+            <tbody></tbody>
+          </table>
       </div>
       <div class="modal-footer">
         <button type="button" class="btn btn-default" data-dismiss="modal">Cancel</button>
-        <button type="button" id="submit-deactivate" class="btn btn-primary"  data-deactivate-url="<%= deactivate_url %>">Delete Applications</button>
-      </div>  
+        <button type="button" id="submit-deactivate" class="btn btn-primary"  data-deactivate-url="<%= deactivate_admin_prospects_url %>">Delete Applications</button>
+      </div>
     </div><!-- /.modal-content -->
   </div><!-- /.modal-dialog -->
 </div><!-- /.modal -->
