+++ app/views/admin/prospects/_filter_modal.html.erb
@@ -7,7 +7,7 @@
       </div>
       <div class="modal-body">
         <div class="container">
-        <%= form_tag prospects_path, method: 'get', id: 'filter-prospects-form' do %>
+        <%= form_tag admin_prospects_path, method: 'get', id: 'filter-prospects-form' do %>
             <div class="row">
               <h4>Contact Information</h4>
               <div class="col-md-2">
