+++ app/views/admin/prospects/edit.html.erb
@@ -1,4 +1,5 @@
 <%= simple_form_for @prospect,  html: { class: 'form-horizontal' },
+  url: admin_prospect_path(@prospect),
   wrapper_mappings: {
     check_boxes: :horizontal_radio_and_checkboxes,
     radio_buttons: :vertical_radio_and_checkboxes,
@@ -6,11 +7,11 @@
     boolean: :horizontal_boolean
   } do |form| %>
   <%= form.error_notification %>
-  <%= render 'id_and_semester_step', form: form %>
-  <%= render 'contact_info_step', form: form %>
-  <%= render 'work_experience_step', form: form %>
-  <%= render 'skills_step', form: form %>
-  <%= render 'availability_step', form: form %>
+  <%= render 'prospects/id_and_semester_step', form: form %>
+  <%= render 'prospects/contact_info_step', form: form %>
+  <%= render 'prospects/work_experience_step', form: form %>
+  <%= render 'prospects/skills_step', form: form %>
+  <%= render 'prospects/availability_step', form: form %>
   <h3>Resume
     <% if @current_user %>
       <% if @current_user.admin %>
@@ -21,7 +22,7 @@
     <% end %>
   </h3>
   <div class="well">
-    <%= link_to 'Back', prospect_path(@prospect), class: 'btn btn-default' %>
+    <%= link_to 'Back', admin_prospect_path(@prospect), class: 'btn btn-default' %>
     <%= form.submit 'Save', class: 'btn btn-success' %>
   </div>
 <% end %>
