+++ app/views/layouts/application.html.erb
@@ -11,7 +11,7 @@
     <li>
       <% if !session[:prospect_params] %>
         <% if authenticated? %>
-          <%= link_to(  prospects_path, id: 'review-applications-link' ) do  %>
+          <%= link_to(  admin_prospects_path, id: 'review-applications-link' ) do  %>
             <b>Review Applications</b>
           <% end %>
         <% end %>
@@ -21,8 +21,8 @@
         <% end %>
       <% end %>
     </li>
-    <li><%= link_to_if( admin?, "<b>List Configuration</b>".html_safe, configuration_path ) {   } %></li>
-    <li><%= link_to_if( admin?, "<b>User Management</b>".html_safe, users_path ) {   } %></li>
+    <li><%= link_to_if( admin?, "<b>List Configuration</b>".html_safe, admin_configuration_path ) {   } %></li>
+    <li><%= link_to_if( admin?, "<b>User Management</b>".html_safe, admin_users_path ) {   } %></li>
     <li>
         <% if @current_user %>
           <%= link_to( "<b>Sign Out</b>".html_safe, logout_url( service: sign_out_url )  )  %>
