+++ app/views/prospects/_skills_step.html.erb
@@ -2,24 +2,24 @@
 <div class='panel panel-default'>
   <div class="panel-heading"><span>Please Select:</span></div>
   <div class="panel-body" id="skills-promoted">
-    <div class='table-responsive'> 
+    <div class='table-responsive'>
       <table class='table table-bordered'>
         <tbody>
           <% Skill.promoted.in_groups_of(3) do |batch| %>
             <tr>
-              <% batch.each do |skill| %> 
+              <% batch.each do |skill| %>
                 <td>
-                  <% unless skill.nil? %> 
+                  <% unless skill.nil? %>
                     <%= form.check_box( :skill_ids,  { :multiple => true }, skill.id, nil  )  %>
                     <%= form.label(  :skills,  skill.name )  %>
                   <% end %>
                 </td>
-              <% end %> 
+              <% end %>
             </tr>
-          <% end %>   
+          <% end %>
         </tbody>
       </table>
-    </div>   
+    </div>
   </div>
 </div>
 
@@ -29,14 +29,15 @@
     <span class='pull-right'>
       <%= link_to_add_association(form, :skills,
                                   { "data-association-insertion-node": "#skills",
-                                    "data-association-insertion-method": 'append'}) do  %>
-        <i class='glyphicon glyphicon-plus-sign'></i>Add Skill 
-      <% end %> 
-    </span>   
+                                    "data-association-insertion-method": 'append',
+                                    partial: 'prospects/skill_fields'}) do  %>
+        <i class='glyphicon glyphicon-plus-sign'></i>Add Skill
+      <% end %>
+    </span>
   </div>
   <div class="panel-body" id="skills">
     <%= form.fields_for :skills, form.object.special_skills do |skill| %>
-        <%= render 'skill_fields', f: skill %>
+        <%= render 'prospects/skill_fields', f: skill %>
     <% end %>
   </div>
 </div>
