+++ app/views/resumes/edit.html.erb
@@ -1,3 +1,3 @@
 <h2>Resume for <%= @resume.prospect.name %></h2>
 <%= render 'form' %>
-<%= link_to "Back", prospect_path(@resume.prospect), class: 'btn btn-default' %>
+<%= link_to "Back", admin_prospect_path(@resume.prospect), class: 'btn btn-default' %>
