+++ test/system/sort_preserves_filter_test.rb
@@ -7,7 +7,7 @@ class SortPreservesFilterTest < ApplicationSystemTestCase
     page.current_window.resize_to(2048, 2048)
 
     User.create(cas_directory_id: "filterer", name: "filterer", admin: false)
-    visit prospects_path
+    visit admin_prospects_path
     fill_in "username", with: "filterer"
     fill_in "password", with: "any password"
     click_button "Login"
